This PR covers performance enhancements for the following queries:

- random components: using the `cytosol` compartment as an example, response time improvement: `~40s` -> `~1s`
- related reactions: using the `transport_reactions` subsystem as an example, response time improvement: from `~6300ms` -> `~1200ms`